requirements.txt
pyyaml>=6.0
pydantic>=2.0
Flask>=3.0
